Çekirdek iki motorlu oldu: PostgreSQL merkez, SQLite yerel
Verinin tamamı bugüne kadar tek makinenin diskindeydi; ikinci bir makine
açıldığında ders programı boş, öğrenci yönetimi hatalı geliyordu. Doğruluk
kaynağı merkezdeki PostgreSQL'e taşınıyor (ADR 0026). Bu commit o taşımanın
depo katmanını kuruyor; kimse henüz Postgres'e bağlanmıyor.

`km_core/store/dialect.py` — SQLite lehçesini PostgreSQL'e çevirir. Kapsam
ölçülerek daraltıldı; tanımadığı yapıyı `UnsupportedDialect` ile REDDEDER,
sessizce yanlış çevirmez. Yanlış çeviri merkezde yanlış veri demektir ve
belirtisi sebebinden çok uzakta çıkar.

`km_core/store/postgres.py` — `Store` ile birebir aynı yüzey. Modüllerin 383
SQL ifadesine ve 51 göç dosyasına DOKUNULMADI (K6).

`km_core/store/bootstrap.py` — şemayı kuran tek yer. İki motor da buradan
geçer; ayrı iki kurulum yolu yazılsaydı biri diğerinden sessizce ayrışır ve
fark ancak üretimde görünürdü. Doğrulandı: aynı kaynaktan iki motorda da
birebir 130 tablo.

`km_core/store/base.py` — `StoreLike` protokolü. Çekirdek artık somut sınıfa
değil sözleşmeye bakıyor; "hangi motor" sorusu tek yerde kaldı.

`PRAGMA table_info` çekirdek göçlerinden çıkarıldı: introspeksiyon motora ait
oldu. Kalsaydı çekirdek göçleri PostgreSQL'de hiç koşamazdı.

Çıktı kaydı (ADR 0019) motora çivilenmişti — kendi `sqlite3` bağlantısını
açıyordu. Yazıcı işlevle bildiriliyor; SQLite'ta senkron yol aynen duruyor,
PostgreSQL'de araya kuyruk giriyor. Kanıtlanmış yolu asenkron yapmak "dosya
yazıldı ama kayıt düşmedi" penceresini yoktan var ederdi.

mypy KATI KİPİ İLK KEZ GERÇEKTEN KOŞUYOR. `--config-file` olmadan mypy
`backend/pyproject.toml`'u okumuyor ve kapı geçmiş görünüyordu; bayrak
eklenince biriken 14 hata çıktı, hepsi kapatıldı. Tip saplamaları da ilan
edildi (K11) — yoksa kapı yalnız bu makinede yeşil olurdu.

4564 test, ruff ve katı mypy yeşil.

+++ backend/src/km_platform/identity_sync/queue.py
@@ -59,7 +59,7 @@ def backoff_seconds(attempts: int) -> float:
     """`attempts` kez denenmiş bir kaydın bir sonraki denemeye kalan süresi."""
     if attempts <= 0:
         return BASE_BACKOFF_SECONDS
-    return min(BASE_BACKOFF_SECONDS * (2 ** (attempts - 1)), MAX_BACKOFF_SECONDS)
+    return float(min(BASE_BACKOFF_SECONDS * (2 ** (attempts - 1)), MAX_BACKOFF_SECONDS))
 
 
 class AuditQueue:
